Atualizando o imc

+++ src/aula03/IfElse.java
@@ -15,18 +15,23 @@ public class IfElse {
 //        }else {
 //            IO.println("Liberado");
 //        }
+
+        //for para pegar os dados
         double nota = 0;
-        for (int i=1; i<=3;i++){
-            IO.println("Escreva a "+i+"° nota");
-            nota += sc.nextInt();
+        for (int i = 1; i <= 3; i++) {
+            IO.println("Escreva a " + i + "° nota");
+            nota += sc.nextDouble();
         }
-        double media = nota/3;
-        if (media<4){
-            IO.println("Sua nota é "+df.format(media)+" e vc está reprovado");
-        } else if (media<=6) {
-            IO.println("Sua nota é "+df.format(media)+" e vc está de recuperação");
-        } else if (media>6) {
-            IO.println("Sua nota é "+df.format(media)+" e vc está aprovado");
+
+        //calculo media
+        double media = nota / 3;
+        //if para saber a media
+        if (media < 4) {
+            IO.println("Sua nota é " + df.format(media) + " e vc está reprovado");
+        } else if (media <= 6) {
+            IO.println("Sua nota é " + df.format(media) + " e vc está de recuperação");
+        } else if (media > 6) {
+            IO.println("Sua nota é " + df.format(media) + " e vc está aprovado");
         }
 
     }
